test callback picklability
